+++ bentoml/yatai/deployment/sagemaker/serve
@@ -37,7 +37,7 @@ def _serve():
     subprocess.check_call(['ln', '-sf', '/dev/stdout', '/var/log/nginx/access.log'])
     subprocess.check_call(['ln', '-sf', '/dev/stderr', '/var/log/nginx/error.log'])
 
-    nginx = subprocess.Popen(['nginx', '-c', '/opt/program/nginx.conf'])
+    nginx = subprocess.Popen(['nginx', '-c', '/bento/nginx.conf'])
     gunicorn_app = subprocess.Popen(
         [
             'gunicorn',
