+++ tests/deployment/sagemaker/test_sagemaker.py
@@ -9,11 +9,11 @@ from botocore.exceptions import ClientError
 import boto3
 from moto import mock_ecr, mock_iam, mock_sts
 
-from bentoml.yatai.deployment.sagemaker import (
+from bentoml.yatai.deployment.sagemaker.operator import (
     _aws_client_error_to_bentoml_exception,
     get_arn_role_from_current_aws_user,
-    SageMakerDeploymentOperator,
 )
+from bentoml.yatai.deployment.sagemaker.operator import SageMakerDeploymentOperator
 from bentoml.yatai.proto.deployment_pb2 import Deployment, DeploymentSpec
 from bentoml.yatai.proto.repository_pb2 import (
     Bento,
@@ -139,9 +139,12 @@ def mock_sagemaker_deployment_wrapper(func):
     @patch('subprocess.check_output', MagicMock())
     @patch('docker.APIClient.build', MagicMock())
     @patch('docker.APIClient.push', MagicMock())
-    @patch('bentoml.yatai.deployment.sagemaker._init_sagemaker_project', MagicMock())
     @patch(
-        'bentoml.yatai.deployment.sagemaker.get_default_aws_region',
+        'bentoml.yatai.deployment.sagemaker.operator._init_sagemaker_project',
+        MagicMock(),
+    )
+    @patch(
+        'bentoml.yatai.deployment.sagemaker.operator.get_default_aws_region',
         MagicMock(return_value='mock_region'),
     )
     def mock_wrapper(*args, **kwargs):
