+++ tests/yatai/test_validators.py
@@ -1,4 +1,4 @@
-from bentoml.yatai.validator import validate_deployment_pb_schema
+from bentoml.yatai.validator import validate_deployment_pb
 from bentoml.yatai.proto.deployment_pb2 import Deployment, DeploymentSpec
 
 
@@ -27,16 +27,16 @@ def _get_test_lambda_deployment_pb():
 
 def test_validate_deployment_pb_schema():
     deployment_pb = _get_test_sagemaker_deployment_pb()
-    assert validate_deployment_pb_schema(deployment_pb) is None
+    assert validate_deployment_pb(deployment_pb) is None
 
     deployment_pb_with_empty_name = _get_test_sagemaker_deployment_pb()
     deployment_pb_with_empty_name.name = ''
-    errors = validate_deployment_pb_schema(deployment_pb_with_empty_name)
+    errors = validate_deployment_pb(deployment_pb_with_empty_name)
     assert errors == {'name': ['required field']}
 
     deployment_pb_with_invalid_service_version = _get_test_sagemaker_deployment_pb()
     deployment_pb_with_invalid_service_version.spec.bento_version = 'latest'
-    errors = validate_deployment_pb_schema(deployment_pb_with_invalid_service_version)
+    errors = validate_deployment_pb(deployment_pb_with_invalid_service_version)
     assert errors['spec'][0]['bento_version'] == [
         'Must use specific "bento_version" in deployment, using "latest" is an '
         'anti-pattern.'
@@ -45,12 +45,12 @@ def test_validate_deployment_pb_schema():
 
 def test_validate_aws_lambda_schema():
     deployment_pb = _get_test_lambda_deployment_pb()
-    assert validate_deployment_pb_schema(deployment_pb) is None
+    assert validate_deployment_pb(deployment_pb) is None
 
     deployment_pb_with_bad_memory_size = _get_test_lambda_deployment_pb()
     deployment_pb_with_bad_memory_size.spec.aws_lambda_operator_config.timeout = 1000
     deployment_pb_with_bad_memory_size.spec.aws_lambda_operator_config.memory_size = 129
-    errors = validate_deployment_pb_schema(deployment_pb_with_bad_memory_size)
+    errors = validate_deployment_pb(deployment_pb_with_bad_memory_size)
     aws_spec_fail_msg = errors['spec'][0]['aws_lambda_operator_config'][0]
 
     assert 'AWS Lambda memory' in aws_spec_fail_msg['memory_size'][0]
